juice_shop/requirements.txt
pytest==7.1.2
selenium==4.4.3
allure-pytest==2.10.0
webdriver-manager==3.8.3